mediacapture-streams/obtaining-local-multimedia-content/navigatorusermedia/deny.html

<!doctype html>
<html>
<head>
  <title>getUserMedia() triggers error callback when auth is denied</title>
<link rel="author" title="Dominique Hazael-Massieux" href="mailto:dom@w3.org"/>
<link rel="help" href="http://w3c.github.io/mediacapture-main/getusermedia.html#error-names">
<link rel="help" href="http://w3c.github.io/mediacapture-main/getusermedia.html#idl-def-MediaStreamError">
<link rel='stylesheet' href='/resources/testharness.css' media='all'/>
</head>
<body>
<p class="instructions">When prompted, <strong>please deny</strong> access to
the video stream.</p>
<h1 class="instructions">Description</h1>
<p class="instructions">This test checks that the error callback is triggered
when user denies access to the video stream.</p>

<div id='log'></div>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src="/common/vendor-prefix.js" data-prefixed-objects='[{"ancestors":["navigator"], "name":"getUserMedia"}]'></script>
<script>
var t = async_test("Tests that the error callback is triggered when permission is denied", {timeout:10000});
t.step(function() {
  navigator.getUserMedia({video: true}, t.step_func(function (stream) {
    assert_unreached("The success callback should not be triggered since access is to be denied");
    t.done();
  }),
  t.step_func(function (error) {
    // NOTE ALEX: last call specs and editor specs differ here:
    // editor draft, 2016: https://w3c.github.io/mediacapture-main/getusermedia.html
    // assert_equals(error.name, "securityError", "securityError returned");
    // last call working draft, 2016: https://www.w3.org/TR/mediacapture-streams/
    // assert_equals(error.name, "PermissionDeniedError", "PermissionDeniedError returned");
    if(error.name != "SecurityError" &&          // FF
       error.name != "PermissionDeniedError") {  // all others
      assert_unreached("Error type is neither securityError or PermissionDeniedError.");
      t.done();
    };
    // NOTE ALEX: this is also nullable, but practically, all the browsers but FF set it to an empty string.
    // assert_equals(error.constraintName, "", "constraintName attribute not set for permission denied");
    if(error.constraintName != "" &&
       error.constraintName != undefined) {
      assert_unreached("constraintName attribute not set for permission denied");
    };
    t.done();
  }));
});
</script>
</body>
</html>
